refactor and expose mining fees
The transaction is fully signed but not published, and returned along
with mining fees. It is up to the caller to call
`electrum.broadcastTransaction(tx)`.
